Creating CRUD endpoints for Invitees  (#619)
* api/serializers/InviteeSerializer.py
Created the InviteeSerializer.py file to implement the serializer for the invitee model. It includes all the fields from the Invitee model.

api/permissions.py
The invitee endpoints should be accessible only by directors, so we added a new permission  CanAccessInvitee in the permissions file to validate that the user has a director role.

api/views/InviteeView.py
Implemented the InviteeViewSet as a ModelViewSet for the invitee. This kind of ViewSet includes implementations for the CRUD operations.
Set the permission classes for the InviteeViewSet as IsAuthenticated and  CanAccessInvitee so just authenticated users who are directors are allowed to access the invitee endpoints.

api/urls.py
Registered the InviteeViewSet with the existing SimpleRouter in the urls.py file, providing the URL prefix ‘invitee’ for the InviteeViewSet.

The unit tests for the permission and the endpoints provided by the InviteeViewSet were also implemented in the following files:
api/tests/test_permissions.py
api/tests/views/test_invitee_view.py

* Added more invitees for testings

* Updated the serializer to add the following fields:

role_name: It's a SerializerMethodField to show the role name
status: It's a calculated field to show if an invitee is Expired, Invited or Resent

Defined the fields we don't need to show in the GET endpoints as write_only

get the role name associated with the invitee given the role id

* Updated the queryset to exclude the accepted invitees

* Updated to add the new fields from the invitee serializer and update the count of the invitees in the fixture we are using for testing

* Created unit test for InviteeSerializer

* Deleted this automigration no longer needed after the invitee model was merged to master

Co-authored-by: Viridiana Ponce Angulo <102187795+viri-ponce@users.noreply.github.com>